# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface Expense {
  id: number;
  amount: number;
  category: string;
  date: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export type ExpenseInputCategory = typeof ExpenseInputCategory[keyof typeof ExpenseInputCategory];


export const ExpenseInputCategory = {
  Food: 'Food',
  Transport: 'Transport',
  'Mobile/Internet': 'Mobile/Internet',
  Entertainment: 'Entertainment',
  Education: 'Education',
  Other: 'Other',
} as const;

export interface ExpenseInput {
  /** @minimum 1 */
  amount: number;
  category: ExpenseInputCategory;
  date: string;
  note?: string;
}

export interface CategorySummary {
  category: string;
  total: number;
}

export interface ExpenseSummary {
  weekTotal: number;
  monthTotal: number;
  byCategory: CategorySummary[];
}

export interface AdviceRequest {
  expenses: Expense[];
}

export interface AdviceResponse {
  advice: string;
}

export interface DeleteResult {
  success: boolean;
}