# lib/api-zod/src/generated/types/adviceRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { Expense } from './expense';

export interface AdviceRequest {
  expenses: Expense[];
}