# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Expense {
  id: number;
  amount: number;
  category: string;
  date: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}