# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { ExpenseInputCategory } from './expenseInputCategory';

export interface ExpenseInput {
  /** @minimum 1 */
  amount: number;
  category: ExpenseInputCategory;
  date: string;
  note?: string;
}