# lib/api-zod/src/generated/types/expenseSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { CategorySummary } from './categorySummary';

export interface ExpenseSummary {
  weekTotal: number;
  monthTotal: number;
  byCategory: CategorySummary[];
}